+++ src/community/components/form/file-upload/file-upload.tsx
@@ -117,7 +117,7 @@ const getUploadErrorHelperText = (rejectedFiles: RejectedFile[], getLabel: ILabe
 };
 
 /**
- * @deprecated Use FileUpload from `@tedi-design-system/react/tedi` instead.
+ * @deprecated Use `FileUpload` from `@tedi-design-system/react/tedi` instead.
  */
 export const FileUpload = (props: FileUploadProps): JSX.Element => {
   const { getLabel } = useLabels();