+++ CHANGELOG.md
@@ -1,3 +1,149 @@
+# [18.1.0-rc.16](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.15...react-18.1.0-rc.16) (2026-07-28)
+
+
+### Bug Fixes
+
+* **community:** add typescript annotations and missing tags to deprecated community components [#563](https://github.com/TEDI-Design-System/react/issues/563) ([#760](https://github.com/TEDI-Design-System/react/issues/760)) ([66466c5](https://github.com/TEDI-Design-System/react/commit/66466c544afc5f8d9be02fcd3fdbf0b3356d5d23))
+
+# [18.1.0-rc.15](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.14...react-18.1.0-rc.15) (2026-07-27)
+
+
+### Bug Fixes
+
+* **LeftPanel:** default added [#746](https://github.com/TEDI-Design-System/react/issues/746) ([afc0acf](https://github.com/TEDI-Design-System/react/commit/afc0acffff0a05974a3e19bc8ec3f368d834e963))
+* **LeftPanel:** expander button aria-controls [#746](https://github.com/TEDI-Design-System/react/issues/746) ([cfe1481](https://github.com/TEDI-Design-System/react/commit/cfe1481b970a1c61fa61c075036c092e3251aaf9))
+* **LeftPanel:** expander button style [#746](https://github.com/TEDI-Design-System/react/issues/746) ([dcecbea](https://github.com/TEDI-Design-System/react/commit/dcecbea1849c0f4991f79be6b2fb07f779bd4c64))
+* **LeftPanel:** focus fix when hideOpenButton [#746](https://github.com/TEDI-Design-System/react/issues/746) ([8b99ad9](https://github.com/TEDI-Design-System/react/commit/8b99ad929c8cadcdb251921cfe4e99e77de0fd18))
+* **LeftPanel:** left-panel component composition [#746](https://github.com/TEDI-Design-System/react/issues/746) ([08fe779](https://github.com/TEDI-Design-System/react/commit/08fe779f04d87ee9be11bc284d290b7a49dc3cce))
+* **LeftPanel:** use ClosingButton and design tokens [#746](https://github.com/TEDI-Design-System/react/issues/746) ([af97cae](https://github.com/TEDI-Design-System/react/commit/af97cae6b71b60c36641214c617cf8d3d49864b9))
+* **LeftPanel:** use design tokens [#746](https://github.com/TEDI-Design-System/react/issues/746) ([2c18efa](https://github.com/TEDI-Design-System/react/commit/2c18efa72af1b5f9c0c40d202194fb57a0e5afa4))
+
+
+### Features
+
+* **left-panel:** Replace LeftPanel Component [#746](https://github.com/TEDI-Design-System/react/issues/746) ([97e71c1](https://github.com/TEDI-Design-System/react/commit/97e71c1654ab6d8a5dda9eef43cf9dfd28b05d47))
+
+# [18.1.0-rc.14](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.13...react-18.1.0-rc.14) (2026-07-20)
+
+
+### Bug Fixes
+
+* **layout:** export ShowAt and HideAt from public entry point [#755](https://github.com/TEDI-Design-System/react/issues/755) ([#756](https://github.com/TEDI-Design-System/react/issues/756)) ([57b8ae2](https://github.com/TEDI-Design-System/react/commit/57b8ae2ff148553850e7d4cd6c7b8a43c3bd7fb4))
+
+# [18.1.0-rc.13](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.12...react-18.1.0-rc.13) (2026-07-20)
+
+
+### Bug Fixes
+
+* **header:** export Header from public entry point [#753](https://github.com/TEDI-Design-System/react/issues/753) ([#754](https://github.com/TEDI-Design-System/react/issues/754)) ([083f9d7](https://github.com/TEDI-Design-System/react/commit/083f9d7c91d6d5af03b40a1fc2fc6837a0dda0d4))
+
+# [18.1.0-rc.12](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.11...react-18.1.0-rc.12) (2026-07-17)
+
+
+### Bug Fixes
+
+* **table:** scrollable table now scrolls top on page change [#750](https://github.com/TEDI-Design-System/react/issues/750) ([#751](https://github.com/TEDI-Design-System/react/issues/751)) ([6154172](https://github.com/TEDI-Design-System/react/commit/6154172ae0f208fd3077d61c9c5f7233036e15c1))
+
+# [18.1.0-rc.11](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.10...react-18.1.0-rc.11) (2026-07-14)
+
+
+### Bug Fixes
+
+* **search:** fix Search component styles [#739](https://github.com/TEDI-Design-System/react/issues/739) ([#745](https://github.com/TEDI-Design-System/react/issues/745)) ([a4571d5](https://github.com/TEDI-Design-System/react/commit/a4571d59b20b6f40ed13ee71418ac82c7c836872))
+
+# [18.1.0-rc.10](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.9...react-18.1.0-rc.10) (2026-07-10)
+
+
+### Bug Fixes
+
+* **map-button:** Added prop functionalities to map-button [#740](https://github.com/TEDI-Design-System/react/issues/740) ([0c0bfef](https://github.com/TEDI-Design-System/react/commit/0c0bfef1a44b21a0fa34dbbed1fe96fcbf909c83))
+* **map-component:** keyboard accessibility fixes + design issues addressed [#740](https://github.com/TEDI-Design-System/react/issues/740) ([cd71fb0](https://github.com/TEDI-Design-System/react/commit/cd71fb06a483947bcfa1f2a94033a25331291d1c))
+* **map-components:** css class naming tweaks [#740](https://github.com/TEDI-Design-System/react/issues/740) ([adfcef7](https://github.com/TEDI-Design-System/react/commit/adfcef79d95da49121e185f155a1a49baa65a47d))
+
+# [18.1.0-rc.9](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.8...react-18.1.0-rc.9) (2026-07-10)
+
+
+### Features
+
+* **popover:** add Header popover variant under Popover [#635](https://github.com/TEDI-Design-System/react/issues/635) ([#726](https://github.com/TEDI-Design-System/react/issues/726)) ([ca42dba](https://github.com/TEDI-Design-System/react/commit/ca42dba2954291c9d4e34dd91751927c0f4e5b73))
+
+# [18.1.0-rc.8](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.7...react-18.1.0-rc.8) (2026-07-10)
+
+
+### Bug Fixes
+
+* **top-nav:** fix nav toggle label styles [#730](https://github.com/TEDI-Design-System/react/issues/730) ([#731](https://github.com/TEDI-Design-System/react/issues/731)) ([73e0928](https://github.com/TEDI-Design-System/react/commit/73e0928b21ad38d6e1cecec214a35b2618cfa6e7))
+
+# [18.1.0-rc.7](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.6...react-18.1.0-rc.7) (2026-07-10)
+
+
+### Features
+
+* **BaseMapSelection:** Add options to display a tooltip [#738](https://github.com/TEDI-Design-System/react/issues/738) ([e3ea862](https://github.com/TEDI-Design-System/react/commit/e3ea8623332084628c138390b859d77187d5f94b))
+* **BaseMapSelection:** adjust border style [#738](https://github.com/TEDI-Design-System/react/issues/738) ([54a33c4](https://github.com/TEDI-Design-System/react/commit/54a33c42339b89b6056105e57abcf91bb2af41c4))
+* **BaseMapSelection:** adjust focus, hover and selected styles [#738](https://github.com/TEDI-Design-System/react/issues/738) ([89f071d](https://github.com/TEDI-Design-System/react/commit/89f071dd22cccfe3046233810315f1ae2eba5dce))
+* **BaseMapSelection:** Use existing element size hook [#738](https://github.com/TEDI-Design-System/react/issues/738) ([0c678ca](https://github.com/TEDI-Design-System/react/commit/0c678ca7c196373c0387e8c9768b8a4a96cdbcc6))
+
+# [18.1.0-rc.6](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.5...react-18.1.0-rc.6) (2026-07-07)
+
+
+### Bug Fixes
+
+* **split-pane:** Added Figma link to storybook [#727](https://github.com/TEDI-Design-System/react/issues/727) ([ff18e02](https://github.com/TEDI-Design-System/react/commit/ff18e0223eae183c88c59e9e42e75e4ee026fefd))
+* **split-pane:** Adjusted comment [#727](https://github.com/TEDI-Design-System/react/issues/727) ([afb73d7](https://github.com/TEDI-Design-System/react/commit/afb73d7c7b50064178301f883e4315deb06cf5f7))
+* **split-pane:** coderabbit feedback fixes [#727](https://github.com/TEDI-Design-System/react/issues/727) ([c53b7d0](https://github.com/TEDI-Design-System/react/commit/c53b7d0b96f370669ac6511716a871c5bdea3c28))
+* **split-pane:** Design tweaks to the SplitPane buttons [#727](https://github.com/TEDI-Design-System/react/issues/727) ([c5f9e3f](https://github.com/TEDI-Design-System/react/commit/c5f9e3f37a231eb82929483b454669949f482756))
+* **split-pane:** min-max ratio props + SB example fix [#727](https://github.com/TEDI-Design-System/react/issues/727) ([02974e8](https://github.com/TEDI-Design-System/react/commit/02974e8187c93a950e70bbb91496bf470074cd30))
+
+
+### Features
+
+* **split-pane:** Added SplitPane component [#727](https://github.com/TEDI-Design-System/react/issues/727) ([edec900](https://github.com/TEDI-Design-System/react/commit/edec900758dff03d7a311bd505fd580d30c0c2dc))
+
+# [18.1.0-rc.5](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.4...react-18.1.0-rc.5) (2026-07-07)
+
+
+### Bug Fixes
+
+* WCAG fixes [#699](https://github.com/TEDI-Design-System/react/issues/699) ([183bbc4](https://github.com/TEDI-Design-System/react/commit/183bbc45b9e276f9805f4e8ade86d4d1c96fa6aa))
+
+
+### Features
+
+* **basemap option:** add disabled state to option [#699](https://github.com/TEDI-Design-System/react/issues/699) ([38937b3](https://github.com/TEDI-Design-System/react/commit/38937b38a183378f92ba22364fb28c6804c6d80b))
+* BaseMapSelection uses popover, handles transparency. Selected works for non historical. [#699](https://github.com/TEDI-Design-System/react/issues/699) ([209535a](https://github.com/TEDI-Design-System/react/commit/209535aca099844d2570a60a57ebe2085b71e420))
+* **basemapselection:** add uncontrolled transparency value [#699](https://github.com/TEDI-Design-System/react/issues/699) ([2cace0e](https://github.com/TEDI-Design-System/react/commit/2cace0e00377cb39159614bba9020f6df4a2576f))
+* **basemapselection:** coderabbit fixes [#699](https://github.com/TEDI-Design-System/react/issues/699) ([a7b14c3](https://github.com/TEDI-Design-System/react/commit/a7b14c3adf7acceef065e29db0c8687cde96bb41))
+* **BaseMapSelection:** review fixes [#699](https://github.com/TEDI-Design-System/react/issues/699) ([216a394](https://github.com/TEDI-Design-System/react/commit/216a39431016cafeb93b49f44f44f4fd1d80bdac))
+
+# [18.1.0-rc.4](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.3...react-18.1.0-rc.4) (2026-07-06)
+
+
+### Features
+
+* **info-button:** focus ring offset reduced [#719](https://github.com/TEDI-Design-System/react/issues/719) ([#724](https://github.com/TEDI-Design-System/react/issues/724)) ([cabb7eb](https://github.com/TEDI-Design-System/react/commit/cabb7eb71e6bf775c0cf51612f90da443d580ae1))
+
+# [18.1.0-rc.3](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.2...react-18.1.0-rc.3) (2026-07-03)
+
+
+### Features
+
+* **label-provider:** turn default locale to "et", remove decorations from stories [#683](https://github.com/TEDI-Design-System/react/issues/683) ([#725](https://github.com/TEDI-Design-System/react/issues/725)) ([fbab5a9](https://github.com/TEDI-Design-System/react/commit/fbab5a989c5dc0d5ae9ae077ada924cc0f00e203))
+
+# [18.1.0-rc.2](https://github.com/TEDI-Design-System/react/compare/react-18.1.0-rc.1...react-18.1.0-rc.2) (2026-07-03)
+
+
+### Bug Fixes
+
+* **button,info-button:** fix neutral button dark mode outline, add InfoButton examples [#694](https://github.com/TEDI-Design-System/react/issues/694) ([#720](https://github.com/TEDI-Design-System/react/issues/720)) ([1f6dc51](https://github.com/TEDI-Design-System/react/commit/1f6dc510067bba1220d4dbc6cf4b859bb150aef5))
+
+# [18.1.0-rc.1](https://github.com/TEDI-Design-System/react/compare/react-18.0.0...react-18.1.0-rc.1) (2026-07-03)
+
+
+### Features
+
+* **header:** add Header top [#489](https://github.com/TEDI-Design-System/react/issues/489) ([#703](https://github.com/TEDI-Design-System/react/issues/703)) ([26e55aa](https://github.com/TEDI-Design-System/react/commit/26e55aa124137fe03a2f524689a6887526383c4a))
+
 # [18.0.0](https://github.com/TEDI-Design-System/react/compare/react-17.1.0...react-18.0.0) (2026-07-02)